Extract Horizontal and Vertical Layout in its own components
